+++ packages/docz-theme-default/src/components/ui/H6.tsx
@@ -1,5 +1,5 @@
 import styled from 'react-emotion'
 
 export const H6 = styled('h6')`
-  ${p => p.theme.styles.h6};
+  ${p => p.theme.mq(p.theme.styles.h6)};
 `