+++ packages/docz-theme-default/src/components/ui/InlineCode.tsx
@@ -1,0 +1,7 @@
+import styled from 'react-emotion'
+
+export const InlineCode = styled('code')`
+  background: ${p => p.theme.colors.codeBg};
+  color: ${p => p.theme.colors.codeColor};
+  ${p => p.theme.mq(p.theme.styles.code)};
+`