+++ packages/docz-theme-default/src/components/ui/Link.tsx
@@ -1,0 +1,14 @@
+import styled from 'react-emotion'
+
+export const Link = styled('a')`
+  &,
+  &:visited,
+  &:active {
+    text-decoration: none;
+    color: ${p => p.theme.colors.link};
+  }
+
+  &:hover {
+    color: ${p => p.theme.colors.link};
+  }
+`