+++ packages/docz-theme-default/src/components/ui/Paragraph.tsx
@@ -1,0 +1,5 @@
+import styled from 'react-emotion'
+
+export const Paragraph = styled('p')`
+  color: ${p => p.theme.colors.text};
+`