+++ packages/docz-theme-default/src/components/ui/Pre.tsx
@@ -7,7 +7,11 @@ import styled, { cx } from 'react-emotion'
 
 const PreStyled = styled('pre')`
   border: 1px solid ${p => p.theme.colors.border};
-  ${p => p.theme.styles.pre};
+  margin: 2em 0;
+  border-radius: 5px;
+  background: ${p => p.theme.colors.preBg};
+  ${p => p.theme.prismTheme};
+  ${p => p.theme.mq(p.theme.styles.pre)};
 `
 
 interface PreProps {
@@ -22,7 +26,6 @@ export class Pre extends PureComponent<PreProps> {
     const { children } = this.props
     const childrenProps = children.props.props
     const childrenClassName = childrenProps && childrenProps.className
-
     const className = cx('react-prism', this.props.className, childrenClassName)
 
     return (