+++ packages/docz-theme-default/src/components/ui/Table.tsx
@@ -12,17 +12,51 @@ export const Table = styled('table')`
   border-style: hidden;
   border-radius: 5px;
   font-size: 14px;
-  color: ${p => p.theme.colors.grayDark};
+  color: ${p => p.theme.colors.tableColor};
+
+  ${p =>
+    p.theme.mq({
+      overflowY: ['hidden', 'hidden', 'hidden', 'initial'],
+      display: ['block', 'block', 'block', 'table'],
+    })};
 
   & thead {
-    color: ${p => p.theme.colors.grayDark};
-    background: ${p => p.theme.colors.grayLight};
+    color: ${p => p.theme.colors.theadColor};
+    background: ${p => p.theme.colors.theadBg};
   }
 
   & thead th {
     text-align: left;
     font-weight: 400;
     padding: 20px 20px;
+
+    &:nth-child(1) {
+      ${p =>
+        p.theme.mq({
+          width: ['20%', '20%', '20%', 'auto'],
+        })};
+    }
+
+    &:nth-child(2) {
+      ${p =>
+        p.theme.mq({
+          width: ['10%', '10%', '10%', 'auto'],
+        })};
+    }
+
+    &:nth-child(3) {
+      ${p =>
+        p.theme.mq({
+          width: ['10%', '10%', '10%', 'auto'],
+        })};
+    }
+
+    &:nth-child(4) {
+      ${p =>
+        p.theme.mq({
+          width: ['20%', '20%', '20%', 'auto'],
+        })};
+    }
   }
 
   & tbody td {
@@ -36,5 +70,5 @@ export const Table = styled('table')`
     border-top: 1px solid ${p => p.theme.colors.border};
   }
 
-  ${p => p.theme.styles.table};
+  ${p => p.theme.mq(p.theme.styles.table)};
 `