+++ packages/docz-theme-default/src/styles/colors.ts
@@ -1,11 +1,10 @@
-export const main = '#272833'
-export const primary = '#0B5FFF'
 export const white = '#FFFFFF'
-export const grayLight = '#F5F7FF'
-export const gray = '#D3DCE6'
-export const grayDark = '#8492A6'
+export const grayExtraLight = '#EEF1F5'
+export const grayLight = '#CED4DE'
+export const gray = '#7D899C'
+export const grayDark = '#2D3747'
+export const grayExtraDark = '#1D2330'
+export const dark = '#13161F'
 
-export const text = main
-export const link = primary
-export const background = white
-export const border = gray
+export const blue = '#0B5FFF'
+export const skyBlue = '#1FB6FF'