+++ packages/docz-theme-default/src/styles/index.ts
@@ -1,32 +1,28 @@
-import * as colors from './colors'
-
 export const styles = {
   body: {
     fontFamily: "'Source Sans Pro', Helvetica, sans-serif",
     fontSize: '16px',
     lineHeight: 1.5,
   },
-  sidebar: {
-    padding: 20,
-    width: 320,
-  },
   container: {
-    padding: '50px 50px 100px',
     width: 960,
+    maxWidth: '100%',
+    padding: ['20px 30px', '50px 50px 100px'],
   },
   h1: {
     margin: '30px 0',
-    fontSize: 48,
+    fontSize: [36, 48],
     fontWeight: 600,
   },
   h2: {
-    margin: '50px 0 20px',
-    fontSize: 32,
+    margin: ['25px 0 20px', '50px 0 20px'],
+    lineHeight: ['1.2em', '1.5em'],
+    fontSize: [30, 32],
     fontWeight: 400,
   },
   h3: {
     margin: '30px 0 20px',
-    fontSize: 24,
+    fontSize: [22, 24],
     fontWeight: 400,
   },
   h4: {
@@ -46,36 +42,22 @@ export const styles = {
     margin: '10px 0 10px 20px',
   },
   playground: {
-    padding: '2rem',
+    padding: ['1.5em', '2em'],
+  },
+  code: {
+    margin: '0 3px',
+    padding: '4px 6px',
+    borderRadius: '3px',
+    fontFamily: "'Inconsolata', monospace",
+    fontSize: 16,
   },
   pre: {
+    padding: ['1.5em', '2em'],
     fontFamily: "'Inconsolata', monospace",
     fontSize: 16,
   },
   table: {
     fontFamily: "'Inconsolata', monospace",
     fontSize: 16,
-  },
-  tooltip: {
-    wrapper: {
-      color: colors.primary,
-    },
-    content: {
-      backgroundColor: colors.main,
-      color: colors.grayLight,
-    },
-    tooltip: {
-      display: 'flex',
-      alignItems: 'center',
-      width: 220,
-      maxWidth: 220,
-      padding: 5,
-      backgroundColor: colors.main,
-      borderRadius: '3px',
-      fontSize: 16,
-    },
-    arrow: {
-      borderTop: `solid ${colors.main} 5px`,
-    },
   },
 }