+++ packages/docz-theme-default/src/styles/prism/index.ts
@@ -1,0 +1,2 @@
+export { theme as dark } from './dark'
+export { theme as light } from './light'